Disable Android automatic text restoration for consumables timers
This appears to use the view's ID as a key - since we had multiple
views with the same ID, the value for that key was always set to
whatever view was last in the hierarchy when the state was saved.

We're doing our own state restoration here anyway, so turning this off
is fine.