Add NotifyService to play notification sounds on timer finished events
Can use this service as a hub to notify the user of events when they're
not looking at their phone.

Realistically, I think this should be proper notifications - playing a
sound with no visual cue or way to figure out where that sound came
from is a pretty poor way of conveying this. I think a custom sound for
each event type would be nice, too, but I suspect that's going to get
complex pretty quickly (depending on where I source the sounds).